refactor(quad,bspline,grid): move _validate_float_dtype to root _array_utils
Moves the generic float-dtype validator out of pantr.basis._basis_utils
into the root-level pantr._array_utils module, which is the correct home
for shared helpers used across top-level modules (quad, change_basis) as
well as subpackages. pantr.basis._basis_utils re-exports it so existing
callers in bspline and change_basis need no import changes.

Also documents that dtype=None is accepted (resolves to float64 per NumPy
semantics) and tightens the _grid_utils module docstring.
